// web/src/api/model/getParents200ItemStudentsItem.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * school-manager-api
 * OpenAPI spec version: 1.0
 */
import type { GetParents200ItemStudentsItemGender } from './getParents200ItemStudentsItemGender';

export type GetParents200ItemStudentsItem = {
    date_of_birth: string;
    gender: GetParents200ItemStudentsItemGender;
    grade: number;
    id: string;
    name: string;
    surname: string;
    teacher_id: string;
};