// web/src/api/model/updateStudentRequest.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * school-manager-api
 * OpenAPI spec version: 1.0
 */
import type { UpdateStudentRequestGender } from './updateStudentRequestGender';

export interface UpdateStudentRequest {
    date_of_birth: string;
    gender: UpdateStudentRequestGender;
    grade: number;
    name: string;
    surname: string;
    teacher_id: string;
}