// web/src/api/model/updateStudentRequestGender.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * school-manager-api
 * OpenAPI spec version: 1.0
 */

export type UpdateStudentRequestGender =
    (typeof UpdateStudentRequestGender)[keyof typeof UpdateStudentRequestGender];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateStudentRequestGender = {
    Male: 'Male',
    Female: 'Female',
} as const;